bug 3794: tree viewer branch/tip label options: Confusing "Minimum size" option should be improved / fixed
tweaked the code a bit to get rid of unnecessary instanceof checks.

I also changed the policy of what happens when the user tries to increase the minimum beyond the current maximum (and vice versa): Previously, the code would just forbid the minimum to become any larger. I've changed it so that it now auto-increments the maximum instead. I think this behaviour is more intuitive and is also the way the Dotplot's window size and stringency behave.

It still resets to the previous value rather than the maximum allowed value if you paste in a very large number, but I think this is indeed a policy somewhere in Java and we can't change it.

Steven to review

git-svn-id: http://svn.code.sf.net/p/jebl/code/trunk@829 6721810a-1124-0410-938e-bb277542c816