Tweak for bug 3637: Nexus: Illegal XML characters in Nexus
Previously only the Nexus importer was fixed, but this can also occur in the Newick importer - in fact, it can occur in any client of ImportedFileOriginalText. These importers import DefaultPhylogenyDocument, DefaultSameTaxaTreesDocument and DefaultTreeDocument. I've made them throw an Exception already when calling setSource() rather than only in toXML(). Because these are in the PublicAPI, I could only let them throw an UncheckedException, so I'm just throwing a org.jdom.IllegalDataException as toXML() normally would.

I've then tweaked the importers to catch these exceptions rather than check the validity of the text first and only then attempting the import, as this means the code that calls the dangerous methods (e.g. setSource()) and the code that handles the error case are close together. Previously NexusDocumentImporter would check the file text for validity, although the code that depended on this was in a totally different place. While Richard's fix for the NExusDocumentImporter was valid, I think this way it's cleaner and more robust against changes in the future (e.g. if NexusDocumentImporter is changed to not use ImportedFileOriginalText any more, if we create a new importer that uses it).

Richard to review

git-svn-id: http://svn.code.sf.net/p/jebl/code/trunk@838 6721810a-1124-0410-938e-bb277542c816