bug 4181: that bloddy cannot compute distance matrix exception thing again (talkback group 254) Matt or Matthew to review
To fix this bug, I proceeded as follows:

1. Temporarily made CannotBuildDistanceMatrixException extend Exception, so it became a checked exception.

2. In all places where it got thrown, I either:
   a) declared the enclosing method to throw it, or
   b) translated it into a DocumentOperationException, or
   c) if it shouldn't occur because the method parameter 
      useTwiceMaximumDistanceWhenPairwiseDistanceNotCalculatable was true,
      I translated it into a RuntimeException

I iterated step 2 until the code compiled. Then, I reverted step 1 and all steps 2 c) (i.e. in places where it shouldn't happen, we now throw the unchecked CannotBuildDistanceMatrixException again). So if we ever observe the exception in talkback again, this should mean that one of the methods which has a useTwiceMaximumDistanceWhenPairwiseDistanceNotCalculatable parameter isn't fulfilling its contract.

git-svn-id: http://svn.code.sf.net/p/jebl/code/trunk@842 6721810a-1124-0410-938e-bb277542c816