bug 4303: Newick: Export sample linnaeus result to newick and it can't import again
Made the newick importer accept trees that have nodes with outdegree 1. In the future, we may decide to automatically replace an internal node with outdegree 1 with its child node, possibly somehow concatenating labels. However, this would have to be done outside the importer, possibly even just when the document is viewed (I think this would make sense e.g. for Linnaeus).

The changes for this are in lines 167-174.

However, I also added some javadoc to the rest of the file. 

Richard to review.

git-svn-id: http://svn.code.sf.net/p/jebl/code/trunk@888 6721810a-1124-0410-938e-bb277542c816