fix failing FastaImporterTest.testInvalidCharacters.
When importing a binary file (with non-ASCII characters) with FASTA nucleotide or amino acid (*not* for autodetect), the FastaImporter would throw a DocumentImportException. However the test depends on the importer silently replacing all invalid characters with '?' when the expected sequence type was known. As this silent replacement is also the current but undocumented policy of jebl Sequences (BasicSequence, CanonicalSequence), I've adjusted the import helper used by the FASTA importer to not throw a DocumentImportException in this case. When encountering non-ASCII characters which it can't store in its internal ByteBuilder, it now switches to using a StringBuilder instead (which may use more memory).

Matt to review

git-svn-id: http://svn.code.sf.net/p/jebl/code/trunk@890 6721810a-1124-0410-938e-bb277542c816