make PaupOperation generate legal tree labels rather than ones that create space. This made PaupIntegrationTest fail.
Also, let the NexusExporter fail if someone tries to create a nexus file with illegal tree names (the proper solution for this would probably be to generate a translation table, but I don't think it's important enough to introduce this support in NexusExporter).

Also, put a missing space in an error message.

Amy has already reviewed my change to PaupOperation. Matt to review the remaining changes.

git-svn-id: http://svn.code.sf.net/p/jebl/code/trunk@909 6721810a-1124-0410-938e-bb277542c816