Futher changes to GEN-3955 : OutOfMemoryError when translating a very large sequence
Bruce's change only reduced the memory usage, but did not address the suggested solution of warning the user if they don't have enough memory to translate, so effectively Bruces's solution just increased the sequence size for which the bug will occur.

So, my changes made it use much less memory (about 10% of what it used before by implementing Tobias's suggestion and some futher changes by me) and it also now warns the user if it has insufficient memory to either translate or to extract a region from a sequence. I was unable to write a useful test case but I have discovered our current free memory checking code is inadequate and I'll be looking into fixing that separately.

Bruce to review.

git-svn-id: http://svn.code.sf.net/p/jebl/code/trunk@912 6721810a-1124-0410-938e-bb277542c816