GEN-8512 - nb this was an assertion not a crash
It happened when the panel width was less than the width of the longest tip label, which caused the tree tips to be behind the origi(o)n.  My fix was to lock the origin and scale to positive vales.  Worst case we lose a bit of the labels off the right hand side of the window, but I think this case is so rare (and happens when you wouldn't be trying to read the panel anyway) that it doesn't matter.

git-svn-id: http://svn.code.sf.net/p/jebl/code/trunk@1018 6721810a-1124-0410-938e-bb277542c816