GEN-25368 Make the tree viewer paint huge trees even faster
Revert my change to getNodeAt which only looked at the visible nodes because it introduced an inconsistency between mousing over and clicking. When you clicked it could select a different node to the one that was highlighted on mouse over. I'm not up for trying to fix that right now.

git-svn-id: http://svn.code.sf.net/p/jebl/code/trunk@1274 6721810a-1124-0410-938e-bb277542c816