GEN-25368 Make the tree viewer paint huge trees even faster
Fix a couple of issues with the "Align Taxon Labels":

- The lines would disappear if you zoomed in such that the branches were not visible due to the clipping we added.
- It would draw all of the lines when zoomed out on a huge tree even though the labels were hidden, resulting in a giant black mess that was slow to draw.

by moving the code for drawing the callout lines next to the label drawing code so they show/hide in tandem.

Also removed an unused local variable and changed the appearance of the callout line to look good when not antialiased.

git-svn-id: http://svn.code.sf.net/p/jebl/code/trunk@1284 6721810a-1124-0410-938e-bb277542c816